+++ src/OpenTelemetry.Exporter.Zipkin/CHANGELOG.md
@@ -7,7 +7,7 @@
   ([#1770](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1770))
 * Removes ability to configure ServiceName for Zipkin. ServiceName must come
   via Resource. If service name is not found in Resource, Zipkin uses
-  GetDefaultResource API from the SDK.
+  GetDefaultResource() from the SDK to obtain it.
   [#1768](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1768)
 
 ## 1.0.0-rc2